requirements.txt
# FastAPI Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Serverless Adapter for Netlify
mangum==0.17.0

# Database Dependencies
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
bcrypt==4.0.1
pyjwt==2.8.0
email-validator==2.1.0

# Astrology & Astronomical Calculations
pyswisseph==2.10.3.2

# Utility Dependencies
requests==2.31.0
pytz==2023.3

# HTTP Client
httpx==0.25.2

# Date/Time Utilities
python-dateutil==2.8.2

# AI/OpenAI Integration
openai==1.3.8